Fix double casting bug and SecureRandom bug
Based on SonarCloud recommendations, cast some values as doubles to prevent type error and use a private variable SecureRandom instance to ensure that the RNG is not crackable.

+++ source/core/src/main/com/csse3200/game/components/items/ItemActions.java
@@ -142,8 +142,8 @@ public class ItemActions extends Component {
     int playerPositionAsIntX = (int)Math.ceil(playerPos.x); 
     int playerPositionAsIntY = (int)Math.ceil(playerPos.y);
     
-    int x = (int)Math.ceil(playerPositionAsIntX + xDelta);
-    int y = (int)Math.ceil(playerPositionAsIntY + yDelta);
+    int x = (int)Math.ceil((double)(playerPositionAsIntX) + xDelta);
+    int y = (int)Math.ceil((double)(playerPositionAsIntY) + yDelta);
     
     return new Vector2(x, y);
   }
